Refactor Request + Private UD
