+++ Sources/SwiftNetKit/NetworkService.swift
@@ -35,7 +35,7 @@ public struct NetworkService: NetworkServiceProtocol {
         self.session = URLSession(configuration: sessionConfiguration)
     }
     
-    private func configureCache(for urlRequest: inout URLRequest, with request: any RequestProtocol) {
+    private func configureCache<T>(for urlRequest: inout URLRequest, with request: Request<T>) {
         if let cacheConfig = request.cacheConfiguration {
             let cache = URLCache(
                 memoryCapacity: cacheConfig.memoryCapacity,
@@ -57,11 +57,11 @@ public struct NetworkService: NetworkServiceProtocol {
         }
     }
     
-    func start<Request: RequestProtocol>(
-        _ request: Request,
+    func start<T>(
+        _ request: Request<T>,
         retries: Int = 0,
         retryInterval: TimeInterval = 1.0
-    ) async throws -> Request.ResponseType {
+    ) async throws -> T {
         var urlRequest = request.buildURLRequest()
         
         CookieManager.shared.includeCookiesIfNeeded(for: &urlRequest, includeCookies: request.includeCookies)
@@ -86,7 +86,7 @@ public struct NetworkService: NetworkServiceProtocol {
                 }
                 
                 do {
-                    let decodedObject = try JSONDecoder().decode(Request.ResponseType.self, from: data)
+                    let decodedObject = try JSONDecoder().decode(T.self, from: data)
                     return decodedObject
                 } catch {
                     throw NetworkError.decodingFailed
@@ -103,11 +103,11 @@ public struct NetworkService: NetworkServiceProtocol {
         throw NetworkError.requestFailed(error: lastError ?? NetworkError.unknown)
     }
     
-    func start<Request: RequestProtocol>(
-        _ request: Request,
+    func start<T>(
+        _ request: Request<T>,
         retries: Int = 0,
         retryInterval: TimeInterval = 1.0,
-        completion: @escaping (Result<Request.ResponseType, Error>) -> Void
+        completion: @escaping (Result<T, Error>) -> Void
     ) {
         var urlRequest = request.buildURLRequest()
         
@@ -145,7 +145,7 @@ public struct NetworkService: NetworkServiceProtocol {
                 
                 if let data = data {
                     do {
-                        let decodedObject = try JSONDecoder().decode(Request.ResponseType.self, from: data)
+                        let decodedObject = try JSONDecoder().decode(T.self, from: data)
                         completion(.success(decodedObject))
                     } catch {
                         completion(.failure(NetworkError.decodingFailed))
